Grad 657 update upgrade plan

+++ course-service/src/main/java/com/example/courseservice/service/CourseService.java
@@ -42,8 +42,8 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import jakarta.transaction.Transactional;
 
-import java.awt.*;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
@@ -71,7 +71,8 @@ public class CourseService {
     private final LessonMapper lessonMapper;
     private final NotificationService notificationService;
     CloudinaryService cloudinaryService;
-
+    TopicRepository topicRepository;
+    CourseSummaryRepository courseSummaryRepository;
 
     public Page<CourseCreationResponse> getAllCourses(
             Boolean isAvailable, Boolean isCompletedCreation,
@@ -107,14 +108,17 @@ public class CourseService {
         Page<Course> courses = courseRepository.findAll(specification, pageable);
 
         return courses.map(course -> {
-            //int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0; //lessonRepository.countByCourse_CourseId(course.getCourseId());
-            List<Section> sections = course.getSections();
-            //int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
-
             AdminCourseCreationResponse response = courseMapper.toAdminCourseCreationResponse(course);
-            //response.setLessonCount(lessonCount);
+
+            int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0; //lessonRepository.countByCourse_CourseId(course.getCourseId());
+            response.setLessonCount(lessonCount);
+
+            List<Section> sections = course.getSections();
             response.setSections(sections);
-            //response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+
+            int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
+            response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+
             return response;
         });
     }
@@ -172,14 +176,17 @@ public class CourseService {
 
         return result.map(
                 course -> {
-                    //int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0; //lessonRepository.countByCourse_CourseId(course.getCourseId());
-                    List<Section> sections = course.getSections();
-                    //int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
-
                     AdminCourseCreationResponse response = courseMapper.toAdminCourseCreationResponse(course);
-                    //response.setLessonCount(lessonCount);
+
+                    int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0; //lessonRepository.countByCourse_CourseId(course.getCourseId());
+                    response.setLessonCount(lessonCount);
+
+                    List<Section> sections = course.getSections();
                     response.setSections(sections);
-                    //response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+
+                    int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
+                    response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+
                     return response;
                 }
         );
@@ -211,22 +218,38 @@ public class CourseService {
         });
     }
 
-    public void deleteCourseById(UUID id, String userUid) {
+    @Transactional
+    public void deleteCourseById(UUID courseId, String userUid) {
         UUID userId = ParseUUID.normalizeUID(userUid);
-        Course course = courseRepository.findByCourseIdAndUserId(id, userId);
+        Course course = courseRepository.findByCourseIdAndUserId(courseId, userId);
+
         if (course == null) {
             throw new AppException(ErrorCode.COURSE_NOT_EXISTED);
         }
+
+        // Check if the course is already enrolled by any user
+        boolean existedUserEnrolled = userCoursesRepository.existsByEnrollId_CourseId(courseId);
+        if (existedUserEnrolled) {
+            throw new AppException(ErrorCode.COURSE_ALREADY_ENROLLED_CANNOT_DELETE);
+        }
+
+        log.info("Deleting course with ID: {}", courseId);
 
         if (course.getCourseImage() != null) {
             cloudinaryService.deleteImage(course.getCourseImage());
         }
 
-        course.setTopic(null);
+        try {
+            courseRepository.delete(course);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException(e);
+        }
+        /*course.setTopic(null);
         course = courseRepository.save(course);
         courseRepository.flush();
 
-        courseRepository.delete(course);
+        courseRepository.delete(course);*/
     }
 
     public CourseCreationResponse createCourse(UUID userUid, CourseCreationRequest request) {
@@ -313,7 +336,22 @@ public class CourseService {
 
         Page<Course> result = courseRepository.findAll(specification, pageable);
 
-        return getCourseSearchResponsesOfAdmin(result);
+        return result.map(course -> {
+            AdminCourseSearchResponse response = courseMapper.toAdminCourseSearchResponse(course);
+
+            List<Section> sections = course.getSections();
+            response.setSections(sections);
+
+            int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0; //lessonRepository.countByCourse_CourseId(course.getCourseId());
+            response.setLessonCount(lessonCount);
+
+            int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
+            response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+
+            return response;
+        });
+
+        //return getCourseSearchResponsesOfAdmin(result);
     }
 
     @NotNull
@@ -361,14 +399,14 @@ public class CourseService {
         return result.map(course -> {
             AdminCourseSearchResponse response = courseMapper.toAdminCourseSearchResponse(course);
 
-            //int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0;  //lessonRepository.countByCourse_CourseId(course.getCourseId());
-            //response.setLessonCount(lessonCount);
+            int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0;  //lessonRepository.countByCourse_CourseId(course.getCourseId());
+            response.setLessonCount(lessonCount);
 
             List<Section> sections = course.getSections();
             response.setSections(sections);
 
-            //int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
-            //response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+            int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
+            response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
 
             return response;
         });
@@ -410,7 +448,22 @@ public class CourseService {
 
         //Page<Course>  courses = courseRepository.findAllByCourseNameContainingIgnoreCaseOrDescriptionContainingIgnoreCase(keyword, keyword, pageable);
 
-        return getCourseSearchResponsesOfAdmin(courses);
+        return courses.map(course -> {
+            AdminCourseSearchResponse response = courseMapper.toAdminCourseSearchResponse(course);
+
+            List<Section> sections = course.getSections();
+            response.setSections(sections);
+
+            int lessonCount = course.getLessons() != null ? course.getLessons().size() : 0; //lessonRepository.countByCourse_CourseId(course.getCourseId());
+            response.setLessonCount(lessonCount);
+
+            int numberOfEnrolledStudents = course.getEnrollCourses() != null ? course.getEnrollCourses().size() : 0;
+            response.setNumberOfEnrolledStudents(numberOfEnrolledStudents);
+
+            return response;
+        });
+
+        //return getCourseSearchResponsesOfAdmin(courses);
     }
 
 
@@ -523,7 +576,16 @@ public class CourseService {
             throw new AppException(ErrorCode.USER_NOT_OWN_COURSE);
         }
 
-        return courseMapper.toAdminCourseCreationResponse(course);
+        AdminCourseCreationResponse response = courseMapper.toAdminCourseCreationResponse(course);
+
+       /* if (course.getCourseSummary() != null && course.getCourseSummary().getSummaryContent()!= null) {
+            CourseSummary courseSummary = course.getCourseSummary();
+            response.setAiSummaryContent(courseSummary.getSummaryContent());
+        }*/
+
+
+
+        return response;
     }
 
 
@@ -1012,8 +1074,7 @@ public class CourseService {
         return null;
     }
 
-    private AuthorCourseResponse toAuthorCourseResponse(Course course)
-    {
+    private AuthorCourseResponse toAuthorCourseResponse(Course course) {
         return AuthorCourseResponse.builder()
                 .courseId(course.getCourseId())
                 .courseName(course.getCourseName())
@@ -1089,6 +1150,15 @@ public class CourseService {
 
         Course savedCourse = courseRepository.save(course);
 
+        CourseSummary courseSummary = CourseSummary.builder()
+                        //.courseId(savedCourse.getCourseId())
+                        .courseName(savedCourse.getCourseName())
+                        .summaryContent(null)
+                        .course(savedCourse)
+                        .build();
+
+        courseSummaryRepository.save(courseSummary);
+
         return courseMapper.toAdminCourseCreationResponse(savedCourse);
    }
 
@@ -1096,27 +1166,45 @@ public class CourseService {
            FinalCourseCreationRequest request, UUID courseId,
            UUID userUuid, String userRole
    ) {
-        if (!userRole.equals(PredefinedRole.admin)) {
-            throw new AppException(ErrorCode.USER_IS_NOT_ADMIN);
-        }
+       if (!userRole.equals(PredefinedRole.admin)) {
+           throw new AppException(ErrorCode.USER_IS_NOT_ADMIN);
+       }
 
-        Course course = courseRepository.findById(courseId)
-                .orElseThrow(() -> new AppException(ErrorCode.COURSE_NOT_EXISTED));
+       Course course = courseRepository.findById(courseId)
+               .orElseThrow(() -> new AppException(ErrorCode.COURSE_NOT_EXISTED));
 
-        course = courseMapper.updateCourse(request, course);
+       course = courseMapper.updateCourse(request, course);
 
-        course.setCurrentCreationStep(3);
-        course.setIsAvailable(false);
+       course.setCurrentCreationStep(3);
+       course.setIsAvailable(false);
 
-        Course savedCourse = courseRepository.save(course);
+       Course savedCourse = courseRepository.save(course);
 
-        return courseMapper.toAdminCourseCreationResponse(savedCourse);
+       CourseSummary courseSummary = null;
+
+       if (request.getAiSummaryContent() != null) {
+           courseSummary = courseSummaryRepository.findById(courseId)
+                   .orElse(
+                           CourseSummary.builder()
+                                   //.courseId(courseId)
+                                   .courseName(savedCourse.getCourseName())
+                                   .summaryContent(request.getAiSummaryContent())
+                                   .course(savedCourse)
+                                   .build());
+
+           courseSummary.setSummaryContent(request.getAiSummaryContent());
+           courseSummaryRepository.save(courseSummary);
+       }
+
+       savedCourse.setCourseSummary(courseSummary);
+
+       return courseMapper.toAdminCourseCreationResponse(savedCourse);
    }
 
    public AdminCourseCreationResponse updateGeneralStepInCourseCreation(
               GeneralCourseCreationRequest request,
               UUID courseId, UUID userUuid, String userRole
-    ) {
+   ) {
        if (!userRole.equals(PredefinedRole.admin)) {
            throw new AppException(ErrorCode.USER_IS_NOT_ADMIN);
        }
@@ -1132,6 +1220,25 @@ public class CourseService {
        course.setCategories(categories);
 
        Course savedCourse = courseRepository.save(course);
+
+       /*CourseSummary courseSummary = savedCourse.getCourseSummary();
+
+       if (courseSummary != null) {
+           courseSummary.setCourseName(savedCourse.getCourseName());
+           courseSummaryRepository.save(courseSummary);
+       }*/
+
+       CourseSummary courseSummary = courseSummaryRepository.findById(courseId)
+               .orElse(
+                       CourseSummary.builder()
+                               .courseName(savedCourse.getCourseName())
+                               .summaryContent(null)
+                               .course(savedCourse)
+                               .build()
+               );
+
+       courseSummaryRepository.save(courseSummary);
+
 
        return courseMapper.toAdminCourseCreationResponse(savedCourse);
    }
@@ -1154,6 +1261,27 @@ public class CourseService {
        course = courseMapper.updateCourse(request, course);
 
        Course savedCourse = courseRepository.save(course);
+
+       //CourseSummary courseSummary = savedCourse.getCourseSummary();
+
+       CourseSummary courseSummary = null;
+
+       if (request.getAiSummaryContent() != null) {
+           courseSummary = courseSummaryRepository.findById(courseId)
+                   .orElse(
+                           CourseSummary.builder()
+                                   //.courseId(courseId)
+                                   .courseName(savedCourse.getCourseName())
+                                   .summaryContent(request.getAiSummaryContent())
+                                   .course(savedCourse)
+                                   .build()
+                   );
+
+           courseSummary.setSummaryContent(request.getAiSummaryContent());
+           courseSummaryRepository.save(courseSummary);
+       }
+
+       savedCourse.setCourseSummary(courseSummary);
 
        return courseMapper.toAdminCourseCreationResponse(savedCourse);
    }
